added aop application error handler

+++ src/main/java/pl/sda/poznan/springmvcintro/controller/ProductController.java
@@ -9,9 +9,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import pl.sda.poznan.springmvcintro.model.Product;
 import pl.sda.poznan.springmvcintro.repository.ProductRepository;
@@ -64,7 +67,7 @@ public class ProductController {
     if (bindingResult.hasErrors()) {
       return ResponseEntity.badRequest().build();
     }
-    // kontynuujemy przetwarzenie - sprawdzamy, czy w bazie danych jest taki produkt
+    // jeśli błędów nie było kontynuujemy przetwarzenie - sprawdzamy, czy w bazie danych jest taki produkt
     Product productFromDb = this.productRepository.findByNameIgnoringCase(product.getName());
     // jeśli repozytorium coś zwróci, to mamy juz taki produkt w bazie
     // wysyłamy status 409 -> CONFLICT
@@ -77,11 +80,48 @@ public class ProductController {
     return ResponseEntity.status(HttpStatus.CREATED).build();
   }
 
+  //todo: napisać aktualizację encji produkt
+  //todo:  sprawdzić czy taki produkt istnieje - zastanowić sie nad statusem jesli nie istnieje
+  //todo:  zwrócić status odpowiedzi
 
+  @PutMapping("/product")
+  public ResponseEntity<Product> updateProduct(@RequestBody @Valid Product product,
+      BindingResult bindingResult) {
+    // sprawdzamy, czy uaktualniona wersja produktu jest zgodna z zasadami walidacyjnymi
+    if (bindingResult.hasErrors()) {
+      // jesli mamy błędy walidacji to zwracamy 400 - BAD REQUEST
+      return ResponseEntity.badRequest().body(product);
+    }
+
+    if (product.getId() == null) {
+      return ResponseEntity.badRequest().body(product);
+    }
+
+    //jesli nie było błędów: pobieramy z bazy danych produkt
+    Optional<Product> optionalProduct = productRepository.findById(product.getId());
+
+    if (optionalProduct.isPresent()) {
+      //metoda save zadziała jak metoda merge entityManagera (jesli jest nadane id to wykona update, jesli bez id to wykona save)
+      productRepository.save(product);
+      return ResponseEntity.ok(product);
+    } else {
+      // nie ma produktu o takim id -> probujemy aktualizowac nie istniejacy produkt - zwracamy blad
+      return ResponseEntity.badRequest().body(product);
+    }
+  }
+
+
+  //todo: zaimplementować usuwanie produktu i zwrocic poprawny status
+  //todo: sprawdź czy istnieje produkt o takim id
   @DeleteMapping("/product/{id}")
-  public String deleteProduct(@PathVariable Long id) {
-
-    this.productRepository.deleteById(id);
-    return "OK";
+  public ResponseEntity<Product> deleteProduct(@PathVariable Long id) {
+    if (id == null) {
+      return ResponseEntity.badRequest().build();
+    }
+    if (this.productRepository.existsById(id)) {
+      this.productRepository.deleteById(id);
+      return ResponseEntity.ok().build();
+    }
+    return ResponseEntity.badRequest().build();
   }
 }
